+++ Nos/Service/NostrIdentifier/NostrIdentifier.swift
@@ -14,9 +14,6 @@ enum NostrIdentifier {
     /// A nostr public key
     case npub(publicKey: RawAuthorID)
 
-    /// A nostr private key
-    case nsec(privateKey: String)
-
     /// A nostr note
     case note(eventID: RawEventID)
 
@@ -37,17 +34,15 @@ enum NostrIdentifier {
     static func decode(bech32String: String) throws -> NostrIdentifier {
         let (humanReadablePart, data) = try Bech32.decode(bech32String)
         switch humanReadablePart {
-        case NostrIdentifierPrefix.publicKey:
+        case Nostr.publicKeyPrefix:
             return try decodeNostrPublicKey(data: data)
-        case NostrIdentifierPrefix.privateKey:
-            return try decodeNostrPrivateKey(data: data)
-        case NostrIdentifierPrefix.note:
+        case Nostr.notePrefix:
             return try decodeNostrNote(data: data)
-        case NostrIdentifierPrefix.profile:
+        case Nostr.profilePrefix:
             return try decodeNostrProfile(data: data)
-        case NostrIdentifierPrefix.event:
+        case Nostr.eventPrefix:
             return try decodeNostrEvent(data: data)
-        case NostrIdentifierPrefix.address:
+        case Nostr.addressPrefix:
             return try decodeNostrAddress(data: data)
         default:
             throw NostrIdentifierError.unknownPrefix
@@ -62,16 +57,6 @@ enum NostrIdentifier {
             throw NostrIdentifierError.unknownFormat
         }
         return .npub(publicKey: publicKey)
-    }
-
-    /// Decodes nsec data into a `NostrIdentifier.nsec`.
-    /// - Parameter data: The encoded nsec data.
-    /// - Returns: The `.nsec` with the private key.
-    private static func decodeNostrPrivateKey(data: Data) throws -> NostrIdentifier {
-        guard let privateKey = SHA256Key.decode(base5: data) else {
-            throw NostrIdentifierError.unknownFormat
-        }
-        return .nsec(privateKey: privateKey)
     }
 
     /// Decodes npub data into a `NostrIdentifier.note`.