+++ Nos/Views/NoteTextViewRepresentable.swift
@@ -165,17 +165,15 @@ struct NoteTextViewRepresentable: UIViewRepresentable {
         ) -> Bool {
             if text.count > 1, let range = Range(nsRange, in: self.text.wrappedValue.attributedString) {
                 do {
-                    let identifier = try NostrIdentifier.decode(bech32String: text)
-                    switch identifier {
-                    case .npub:
+                    let (humanReadablePart, _) = try Bech32.decode(text)
+                    if humanReadablePart == Nostr.publicKeyPrefix {
                         self.text.wrappedValue.insertMention(npub: text, at: range)
                         return false
-                    case .note:
+                    } else if humanReadablePart == Nostr.notePrefix {
                         self.text.wrappedValue.insertMention(note: text, at: range)
                         return false
-                    default:
-                        return true
                     }
+                    return true
                 } catch {
                     return true
                 }